Version 3.2 Build 1

+++ Buckets/Model/LeagueStandings/Standing.swift
@@ -16,7 +16,7 @@ struct Standing: Codable, Hashable {
     
     let Wins: Int?
     let Losses: Int?
-    let Percentage: Float?
+    let Percentage: Double?
     
     let ConferenceRecord: String?
     let DivisionRecord: String?
